+++ ui/src/components/app-sidebar.test.ts
@@ -19,6 +19,7 @@ import "../test-helpers/app-sidebar-cases/child-sessions.ts";
 import "../test-helpers/app-sidebar-cases/group-mutations.ts";
 import "../test-helpers/app-sidebar-cases/identity-menu.ts";
 import "../test-helpers/app-sidebar-cases/interactions.ts";
+import "../test-helpers/app-sidebar-cases/new-group-dialog.ts";
 import "../test-helpers/app-sidebar-cases/narration.ts";
 import "../test-helpers/app-sidebar-cases/outbox-badges.ts";
 import "../test-helpers/app-sidebar-cases/pull-request-state.ts";
